Add files via upload
This project is a sleek and modern product landing page designed to showcase a product using HTML and CSS. The landing page aims to captivate visitors with a clean layout, engaging visuals, and compelling content. HTML5: For structuring the content of the landing page.
CSS3: For styling the layout, typography, and visual effects.